requirements.txt
praw
psaw
python-dotenv